requirements.txt
Flask==2.0.3
psycopg2_binary==2.9.1
simplejson==3.17.3
chardet==4.0.0
Jinja2==3.0.1
unicodecsv==0.14.1
Dozer==0.8
numpy==1.22.4
timeout_decorator==0.5.0
Werkzeug==2.0.1
Babel==2.9.1
boto3==1.18.22
pandas==1.4.2
shortuuid==1.0.1
SQLAlchemy==1.4.22
sqlalchemy-redshift==0.8.4
Flask_JWT_Extended==4.2.3
requests==2.25.1
pyinstrument==3.1.3
cached_property==1.5.2
flask_compress==1.10.1
flask_debugtoolbar==0.11.0
flask_sqlalchemy==2.5.1
heroku3==4.2.3
kids==1.0.4
openpyxl==3.0.7
pyexcel==0.6.6
pyexcel-xls==0.6.2
pyexcel-xlsx==0.6.0
python_dateutil==2.8.2
scipy==1.8.1
sendgrid==6.8.0
Unidecode==1.2.0
whoosh==2.7.4
kids.cache==0.0.7
python-dotenv==0.19.0
gunicorn==20.1.0
gspread==5.2.0
hubspot3==3.2.51
celery==5.2.6
redis==4.2.2
httpx==0.23.0
